Refactoring work items
