+++ src/mcp_azure_devops/features/work_items/tools/create.py
@@ -1,0 +1,17 @@
+"""
+Create operations for Azure DevOps work items.
+
+This module provides MCP tools for creating work items.
+"""
+from typing import Optional, Dict, Any
+from mcp_azure_devops.features.work_items.common import get_work_item_client, AzureDevOpsClientError
+
+def register_tools(mcp) -> None:
+    """
+    Register work item creation tools with the MCP server.
+    
+    Args:
+        mcp: The FastMCP server instance
+    """
+    # Placeholder for future implementation
+    pass
