+++ tests/features/work_items/test_tools.py
@@ -1,11 +1,9 @@
 import pytest
 from unittest.mock import MagicMock, patch
-from azure.devops.v7_1.work_item_tracking.models import WorkItem, WorkItemReference, Wiql, ReferenceLinks
-from mcp_azure_devops.features.work_items.tools import (
-    _query_work_items_impl,
-    _get_work_item_impl,
-    _get_work_item_comments_impl
-)
+from azure.devops.v7_1.work_item_tracking.models import WorkItem, WorkItemReference
+from mcp_azure_devops.features.work_items.tools.query import _query_work_items_impl
+from mcp_azure_devops.features.work_items.tools.read import _get_work_item_impl
+from mcp_azure_devops.features.work_items.tools.comments import _get_work_item_comments_impl
 
 # Tests for _query_work_items_impl
 def test_query_work_items_impl_no_results():
@@ -77,7 +75,7 @@ def test_get_work_item_impl_basic():
     }
     mock_client.get_work_item.return_value = mock_work_item
     
-    result = _get_work_item_impl(123, mock_client, detailed=False)
+    result = _get_work_item_impl(123, mock_client)
     
     # Check that the result contains expected basic info
     assert "# Work Item 123: Test Bug" in result
@@ -107,7 +105,7 @@ def test_get_work_item_impl_detailed():
     }
     mock_client.get_work_item.return_value = mock_work_item
     
-    result = _get_work_item_impl(123, mock_client, detailed=True)
+    result = _get_work_item_impl(123, mock_client)
     
     # Check that the result contains both basic and detailed info
     assert "# Work Item 123: Test Bug" in result
@@ -123,7 +121,7 @@ def test_get_work_item_impl_error():
     mock_client = MagicMock()
     mock_client.get_work_item.side_effect = Exception("Test error")
     
-    result = _get_work_item_impl(123, mock_client, detailed=False)
+    result = _get_work_item_impl(123, mock_client)
     
     assert "Error retrieving work item 123: Test error" in result
